- Oprava, pridane DTO

+++ src/main/java/com/example/tbd/JwtTokenUtil.java
@@ -13,10 +13,10 @@ public class JwtTokenUtil {
     private static final String SECRET_KEY = "verysecuresecretkeywith256bits1234567890";
     private final Key key = Keys.hmacShaKeyFor(SECRET_KEY.getBytes());
 
-    public String generateToken(String subject, int customerId, String customerEmail) {
+    public String generateToken(String subject, Long customerId, String customerEmail) {
         return Jwts.builder()
                 .setSubject(subject)
-                .claim("customerId", customerId)
+                .claim("customerId", customerId) // Long hodnota
                 .claim("customerEmail", customerEmail)
                 .setIssuedAt(new Date())
                 .setExpiration(new Date(System.currentTimeMillis() + 86400000)) // Token valid for 24 hours
