+++ src/main/java/com/example/tbd/customer/CustomerRepository.java
@@ -9,11 +9,12 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
     public List<Customer> findAllByName(String name);
     public List<Customer> findAllBySurname(String surname);
     Optional<Customer> findByEmail(String email);
-    boolean existsById(Integer id); // Checks if a customer exists by their ID
+    boolean existsById(Long id); // Checks if a customer exists by their ID
     @Query("SELECT COUNT(c) FROM Customer c WHERE c.roleId = 1")
     @Transactional(readOnly = true)
     long countUsersWithRoleUser();
