+++ src/main/java/com/example/tbd/vehicle/VehicleRepository.java
@@ -11,13 +11,13 @@ import java.util.Optional;
 
 public interface VehicleRepository extends JpaRepository<Vehicle, Integer> {
     List<Vehicle> findAll(); // Find all vehicles
-    Optional<Vehicle> findById(Integer id); // Find vehicle by ID
+    Optional<Vehicle> findById(Long id); // Find vehicle by ID
     Optional<Vehicle> findByVin(String vin);   // Find vehicle by VIN
     Optional<Vehicle> findByPlateNo(String plateNo);
-    List<Vehicle> findByCustomerId(Integer customerId); // Find vehicles by customer ID
+    List<Vehicle> findByCustomerId(Long customerId); // Find vehicles by customer ID
     boolean existsByPlateNo(String plateNo);
     boolean existsByVin(String vin);
-    List<Vehicle> findByCustomerIdAndDeleted(Integer customerId, String deleted);
+    List<Vehicle> findByCustomerIdAndDeleted(Long customerId, String deleted);
     boolean existsByPlateNoAndDeleted(String plateNo, String deleted);
     @Query(value = "SELECT COUNT(*) FROM vehicle WHERE deleted = 'N'", nativeQuery = true)
     long countVehicles();
